[readability] move logic to params_extractor
